requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
python-dotenv==1.0.1
faiss-cpu==1.8.0
sentence-transformers==3.0.1
google-generativeai==0.7.2
pandas==2.2.2
numpy==1.26.4
rank-bm25==0.2.2
python-multipart==0.0.9
httpx==0.27.0
tqdm==4.66.4
pydantic==2.7.1
